+++ src/styles/components/Header.styled.tsx
@@ -20,14 +20,10 @@ const StyledHeader = styled.header`
 
   nav a {
     text-decoration: none;
-    color: #CCC;
+    color: white;
     padding-left: 0.75rem;
     padding-right: 0.75rem;
     font-size: 0.9rem;
-  }
-
-  nav a:hover {
-    color: #FFF;
   }
 `;